Make dimension guardrail bucket drop configurable in ValueMetric
Change the default behaviour of ValueMetric to not drop the bucket when
dimension guardrail is hit. Dropping the entire bucket is too aggressive
and reduces coverage. Add a boolean field to ValueMetric config to
force the bucket to be dropped when dimension guardrail is hit.

If the bucket is not dropped and use_zero_default_base is true, don't
compute diffs from zero default value and only compute diffs from known
bases. It is possible a base might be missing because we didn't record
it once the guardrail was hit rather than it not being preset in the
data. Err on the side of incomplete data rather than incorrect data.

Bug: 408060408
Flag: com.android.os.statsd.flags.keep_value_metric_max_dimension_bucket
Test: statsd_test
Change-Id: I0c43f7fc2392a252c379c62e57233fb17cd4796a