Make anomaly detection work with non-integer types
This CL doesn't change existing anomaly detection logic, just extending
the support float type. We need to revisit the anomaly detection logic
for gauge metrics.

Alert config uses double for threshold, so we use double in statsd for
anomaly tracking. There can be precision loss in converting to double:

[int64 fields] Double type covers the range of int64, and preserves the
precision up to 2^53. So it should be enough.
[float fields] When float fields are used for anomaly detection, they
are converted to double, there may be precision loss.

Test: atest statsd_test
Flag: EXEMPT test only
Bug: 431086298

Change-Id: I6d23c4d7851dac085c309541a19ec0cfca330bca